+++ src/components/search/SearchPosts.tsx
@@ -1,0 +1,3 @@
+export const SearchPosts = () => {
+  return <div></div>;
+};
