+++ src/pages/feed/index.tsx
@@ -10,25 +10,14 @@ import { FeedTagArea } from '../../components/feed/FeedTagArea';
 import { FloatSearch } from '../../components/feed/FloatSearh';
 import { NoPost } from '../../components/feed/NoPost';
 import { Posts, SinglePostResponse } from '../../components/feed/Posts';
+import { useFeed } from '../../hooks/useFeed';
 import { useToast } from '../../hooks/useToats';
 import { userState } from '../../store/Auth/userState';
 import { getCookie } from '../../utils/cookie';
 
 const Feed = () => {
-  const [posts, setPosts] = useState<SinglePostResponse[]>([]);
   const { openToast } = useToast();
-
-  const getFeed = async () => {
-    const feed: PostResponse = await postAPI.ALL(getCookie());
-    if (feed && getCookie()) {
-      console.log(feed);
-    }
-    setPosts(feed.posts);
-  };
-
-  useEffect(() => {
-    getFeed();
-  }, []);
+  const { posts } = useFeed();
 
   return (
     <div>
